refactor(core): split authz grant.rs into model / repository / use_case
authz/grant.rs had grown to ~1040 lines mixing four concerns: the grant
model + role catalog, the GrantRepository port, and the GrantUseCases use
case (plus its tests). Split it into a `grant/` module matching the
use_case.rs / repository.rs layout every sibling application module uses:

- grant/mod.rs      — model + catalog (Scope, Principal, Grant, the
                      GrantableRole catalog, validate_* helpers,
                      is_owner_role, removal_orphans_scope)
- grant/repository.rs — the GrantRepository port
- grant/use_case.rs   — GrantUseCases + its impl + tests

All public paths (authz::grant::{...}, re-exported by authz/mod.rs and
application/mod.rs) are unchanged, so no code outside the module moves.
No behaviour change. Full workspace suite green, clippy clean.
